Move Crossvalidation into BaseNNEstimator
It was previously in BaseNNMixture, but the function generalizes
to all estimators using Neural Nets, not just mixtures

Resolve problem with fit_by_cv assuming an old version of sklearn (0.19)
